test: refactor tenure change tests
